requirements.txt
discord.py>=2.0.1
aiohttp>=3.7.4,<4
Pillow==9.4.0
asyncpg==0.27.0
requests~=2.28.1
colorama==0.4.6
protobuf